pyproject.toml
[tool.poetry]
name = "question-generator-ai"
version = "0.1.0"
description = ""
authors = ["Ashwin M <thatsashwin@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
firecrawl-py = "^1.3.1"
openai = "^1.52.0"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
